Add first attempt of BT Hello World
Managed to get the build working, but the binary doesn't seem to be
correct. Doing a `ldd` in Kindle shows this:
```
linux-vdso.so.1 (0xbee6f000)
/usr/lib/libenvload.so (0xb6f93000)
/home/sighery/x-tools/arm-kindlehf-linux-gnueabihf/arm-kindlehf-linux-gnueabihf/sysroot/usr/lib/libace_bt.so
=> not found
libgcc_s.so.1 => /lib/libgcc_s.so.1 (0xb6f5e000)
libc.so.6 => /lib/libc.so.6 (0xb6e2b000)
/lib/ld-linux-armhf.so.3 (0xb6f9c000)
```

Which obviously the `libace_bt.so` is not in the correct path for that
system.

Was able to get it working for now by manually patching the path:
```
patchelf --replace-needed /bad-path/libace_bt.so libace_bt.so binary
```