Automate patching libace_bt hard link
Requires `patchelf`.

This doesn't actually fix the root issue. Seems that Meson won't play
nice with libraries without `soname`:
https://mesonbuild.com/Shipping-prebuilt-binaries-as-wraps.html#note-for-linux-libraries
https://web.archive.org/web/20250625181332/https://mesonbuild.com/Shipping-prebuilt-binaries-as-wraps.html#note-for-linux-libraries

We don't control these libraries or how they are compiled by Amazon,
and Meson seems unwilling to change this, based on this issue being
opened since 2020:
https://github.com/mesonbuild/meson/issues/7766

We might be able to do something at the `kindle-sdk` level to generate
the `soname`. But for now, this hacky solution after the compilation
step works.