Add first version of kindlebt library
This will be an open source wrapper over `ace_bt`, that can be linked in
other languages, as well as hiding some of the complexity the `ace_bt`
library has.

The project has also been switched from C++ to C, as I realised the
native Golang integration CGO only works with C, not C++. If I was using
C++, I'd need Swig, another third-party tool, to make it work. Most of
the code I wrote already to try out `ace_bt` seemed to have been C
anyway, so I just decided to switch early.

There is a binary being built as well, `kindlebt_test`, but this is just
a simple binary for testing, that will likely be removed at some point.