Fix waitForCondition flag pass by value
Didn't realise that passing a `bool` would be a pass by value, meaning I
was just lucky sometimes and the program was so quick that the variable
would be set to `true` before passing it to `waitForCondition`.

Now that it's an explicit pass by reference, the value will always be up
to date with the value the variable has.