Add get GATT DB functionality
This is a prerequisite step for being able to work with BLE
read/write/notification, as it all then uses this local pGgatt_service,
which has structs for for reading/writing data for each characteristic.

That means `bleGetDatabase` needs to be called before trying to do any
read/write operations.