Add application callbacks functionality
For some callbacks, we'll need to have the application receive them,
like notifications. Even if I wanted to make it all sync, there's just
no way to do so with something like notifications, where the timing of
new messages depends entirely on the BLE server. Additionally, I think
some longer messages might get split into multiple read callbacks?

It gets even trickier because in some cases, we still need the library
code to have access to the callbacks, as it needs to set up code, like
for the GATT DB call, where it requires the callback to be able to make
the local copy of the BLE server's GATT DB.

For now we'll just have wrapper functions and a second callbacks struct,
and just call both library + application callbacks as needed.