Disable library BLE read/write/notify callbacks
The library doesn't care for the read/write/notify callbacks, it was
just at the beginning where we needed to test everything was working as
expected. Now that we know it is, and we know we can hook Go callbacks
instead, the C callbacks just take up CPU cycles and waste time.