Reverse bleRegisterGattClient for pre 5.17
The GATT Client APIs seem to only be included in `libace_bt.so` starting
from 5.17 firmwares. The GATTC functions are however built on older
common APIs that are still part of the ace_bt ABI. That being said,
unfortunately, for some of these older functions we just have no
documentation whatsoever. They are missing from the headers, and all the
binaries are stripped. So reversing the new GATT Client APIs require a
combination of static and dynamic analysis. As of this commit,
`aceBt_bleRegisterGattClient` has been successfully reversed.