Reverse aceAipc_parameter_t struct
Still a lot remains to be figured out about this struct, and we might
never fully map it out, but for now we have a basic mapping that exposes
`function_id`, `buffer` and `callback_id`, seemingly in the correct
places, with the correct types, so it's good enough for our usecase
which is the `gattc_cb_handler` that gets passed this struct and uses
those fields.

This still requires more testing and more code, and actually getting to
use/call the GATTC callbacks, which so far we haven't even gotten to
yet.