Fix GATTC ipc_evt_enum_t values
It seems at some point these values might have changed? Down on the
50-60 range, where ACE_BT_CALLBACK_CLE_* are, these seem to have
remained as is. My decompiled code was showing me the proper enum values
being used there. Likewise, for the `gatts_cb_handler` function, it
showed the proper GATTS values being used, which are in the 70-80 range.

However, for some reason, for the `gattc_cb_handler`, it was using the
90-100 range, which fell on`ACE_BT_CALLBACK_PM_*` ("BT Policy Manager
Callback Interfaces).

This made no sense, and with this mapping, the `gattc_cb_handler` code
would have called a bunch of unmapped Policy Manager values. However, if
instead you move the PM values down to the 80-90 range, and the GATTC up
into 90-100, then it works out perfect, and all the cases in the
decompiled code seem to match up to the proper enum values.