Reverse bleGetDatabase for pre 5.17 firmwares
Yet another GATT Client function that needed reversing for pre 5.17
firmwares. This function call is essentially a requirement for later
read/write operations, as this fills a local struct that will be used
later on for the read/write operations, as well as checks for a given
characteristic ID.

Thankfully reversing this one was fairly easy, and this also helped test
our `pre5170_gattc_cb_handler` event handler, and yep, all working as
expected!