Reverse bleReadCharacteristic for <5.17 firmwares
Like all the previous GATT Client API operations, these symbols are not
in the older libace_bt.so versions.

I'm not 100% confident on the implementation in the
`pre5170_gattc_cb_handler`, as the decompiled code does a ton more
assignments into this struct. So I might have half this struct
empty/unset. But currently I'm just using the `blobValue` data, so it's
working so far. Also printing structs in C is a pain, so I haven't
bothered. Once/if we need the rest of the data, we'll get to it.