Add shim functions for new GATTC APIs
In retrospect, calling the new GATT Client APIs was pretty silly. Of
course, on SF builds, where those symbols are not available, the linker
will fail, since the symbols cannot be found.

So similar to the `getSessionFor*` functions, I'll also need to do this
for all the GATT Client APIs.