Implement bleDiscoverAllServices (all firmwares)
I ran into this issue when comparing the C and Golang implementations,
it seems right after `bleConnect`, if you wait for a second, the
GATT services become available and you can retrieve them with
`bleGetDatabase`/`bleGetService`. I'm still not quite sure, but perhaps
this new `bleDiscoverAllServices` is a necessary step? To first discover
the services, then later on retrieve them?

So far seems to work with the wait removed, so good enough.