Implement bleWriteDescriptor
I hadn't had implemented anything to do with descriptors before as I
didn't need it, or so I thought.

Turns out enabling/disabling notifications, actually goes into the
characteristic descriptors, find a specific descriptor, the
Characteristic Client Configuration Descriptor, and then set a value on
its `BlobValue` data field.

If you have access to the new APIs, when you call `bleSetNotification`,
it's invisibly writing the descriptor for you to enable/disable
notifications on that characteristic. Same thing for the
`bleSetIndication` function, which also uses CCCD but writes 0x0002
instead of 0x0001.

That means that for me to reverse `bleSetNotification`, I needed to
reverse `bleWriteDescriptor` first. This is that reversed
implementation. With both pre and post 5.17 firmware support.