Reverse bleSetNotification for <5.17 firmwares
Like all the other operations, this also needed reversing for pre 5.17
firmwares since the GATT Client APIs are not included. The extra kink
with this API is that it needs to write to a Characteristic Descriptor,
so it needed to be reversed _after_ we had `bleWriteDescriptor` working.

This is also a required step. Even if you know how to write the
Descriptor yourself, and what value to set where, to enable
notifications on a given Characteristic, it seems that will enable the
notification at the controller layer, but not have any of the registered
events in `btmanagerd`, so the event handler will never be called for
those events, even as I see in the logs we're receiving notification
packets.