Add initial kindlebt documentation
Using Doxygen and a bunch of extra Markdown files on top of the main
`kindlebt.h` and `kindlebt_defines.h` documentation. This is just the
initial documentation as there is a lot more that could and should be
documented, but I feel this is a good start to continue building on.